pyproject.toml
[tool.poetry]
name = "pycon"
version = "0.1.0"
description = ""
authors = ["Chakrit Wichian <service@chakrit.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
anyio = "^4.1.0"
dagger-io = "^0.9.4"
django = "^5.0"
psycopg = "^3.1.14"
django-environ = "^0.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
